PS2 Keyboard Test
Added PS2 keyboard test, fixed some definitions, and tweaked splash screen to work on the K

+++ src/include/A2560K/VICKYIII_a2560k.h
@@ -9,7 +9,7 @@
  * Screen Channel A
  */
 
-#define MasterControlReg_A		    ((volatile int *)0xFEC40000)
+#define MasterControlReg_A		    ((volatile unsigned long *)0xFEC40000)
 #define VKY3_MCR_TEXT_EN            0x00000001  /* Text Mode Enable */
 #define VKY3_MCR_TEXT_OVRLY         0x00000002  /* Text Mode overlay */
 #define VKY3_MCR_GRAPH_EN           0x00000004  /* Graphic Mode Enable */
@@ -32,34 +32,34 @@
 #define VKY3_DIP_HIRES              0x4000      /* DIP switch for high resolution mode */
 #define VKY3_PLL_ACTIVE_CLK         0x8000      /* Active Clock --- 0: 25.175Mhz, 1: 40Mhz */
 
-#define BorderControlReg_L_A	    ((volatile int *)0xFEC40004)
+#define BorderControlReg_L_A	    ((volatile unsigned long *)0xFEC40004)
 #define VKY3_BRDR_EN                0x00000001  /* Border Enable */
 #define VKY3_X_SCROLL_MASK          0x00000070  /* X Scroll */
 #define VKY3_X_SIZE_MASK            0x00003f00  /* X Size */
 #define VKY3_Y_SIZE_MASK            0x003f0000  /* Y Size */
 
-#define BorderControlReg_H_A 	    ((volatile int *)0xFEC40008)
-#define BackGroundControlReg_A	    ((volatile int *)0xFEC4000C)
-#define CursorControlReg_L_A	    ((volatile int *)0xFEC40010)
-#define CursorControlReg_H_A	    ((volatile int *)0xFEC40014)
+#define BorderControlReg_H_A 	    ((volatile unsigned long *)0xFEC40008)
+#define BackGroundControlReg_A	    ((volatile unsigned long *)0xFEC4000C)
+#define CursorControlReg_L_A	    ((volatile unsigned long *)0xFEC40010)
+#define CursorControlReg_H_A	    ((volatile unsigned long *)0xFEC40014)
 
-#define LineInterrupt0_A		    ((volatile uint16_t *)0xFEC40018)
-#define LineInterrupt1_A		    ((volatile uint16_t *)0xFEC4001A)
-#define LineInterrupt2_A		    ((volatile uint16_t *)0xFEC4001C)
-#define LineInterrupt3_A		    ((volatile uint16_t *)0xFEC4001E)
+#define LineInterrupt0_A		    ((volatile unsigned long *)0xFEC40018)
+#define LineInterrupt1_A		    ((volatile unsigned long *)0xFEC4001A)
+#define LineInterrupt2_A		    ((volatile unsigned long *)0xFEC4001C)
+#define LineInterrupt3_A		    ((volatile unsigned long *)0xFEC4001E)
 
-#define FONT_Size_Ctrl_A            ((volatile int *)0xFEC40020)
-#define FONT_Count_Ctrl_A           ((volatile int *)0xFEC40024)
+#define FONT_Size_Ctrl_A            ((volatile unsigned long *)0xFEC40020)
+#define FONT_Count_Ctrl_A           ((volatile unsigned long *)0xFEC40024)
 
-#define MousePointer_Mem_A		    ((volatile uint16_t *)0xFEC40400)
-#define MousePtr_A_CTRL_Reg		    ((volatile uint16_t *)0xFEC40C00)
+#define MousePointer_Mem_A		    ((volatile unsigned long *)0xFEC40400)
+#define MousePtr_A_CTRL_Reg		    ((volatile unsigned long *)0xFEC40C00)
 #define MousePtr_En                 0x0001
 
-#define MousePtr_A_X_Pos		    ((volatile uint16_t *)0xFEC40C02)
-#define MousePtr_A_Y_Pos		    ((volatile uint16_t *)0xFEC40C04)
-#define MousePtr_A_Mouse0		    ((volatile uint16_t *)0xFEC40C0A)
-#define MousePtr_A_Mouse1		    ((volatile uint16_t *)0xFEC40C0C)
-#define MousePtr_A_Mouse2		    ((volatile uint16_t *)0xFEC40C0E)
+#define MousePtr_A_X_Pos		    ((volatile unsigned long *)0xFEC40C02)
+#define MousePtr_A_Y_Pos		    ((volatile unsigned long *)0xFEC40C04)
+#define MousePtr_A_Mouse0		    ((volatile unsigned long *)0xFEC40C0A)
+#define MousePtr_A_Mouse1		    ((volatile unsigned long *)0xFEC40C0C)
+#define MousePtr_A_Mouse2		    ((volatile unsigned long *)0xFEC40C0E)
 
 #define ScreenText_A			    ((volatile char *)0xFEC60000)             /* Text matrix */
 #define ColorText_A				    ((volatile uint8_t *)0xFEC68000)	      /* Color matrix */
@@ -70,39 +70,39 @@
  * Screen Channel B
  */
 
-#define MasterControlReg_B		    ((volatile int *)0xFEC80000)
-#define BorderControlReg_L_B	    ((volatile int *)0xFEC80004)
-#define BorderControlReg_H_B 	    ((volatile int *)0xFEC80008)
-#define BackGroundControlReg_B	    ((volatile int *)0xFEC8000C)
-#define CursorControlReg_L_B	    ((volatile int *)0xFEC80010)
-#define CursorControlReg_H_B	    ((volatile int *)0xFEC80014)
+#define MasterControlReg_B		    ((volatile unsigned long *)0xFEC80000)
+#define BorderControlReg_L_B	    ((volatile unsigned long *)0xFEC80004)
+#define BorderControlReg_H_B 	    ((volatile unsigned long *)0xFEC80008)
+#define BackGroundControlReg_B	    ((volatile unsigned long *)0xFEC8000C)
+#define CursorControlReg_L_B	    ((volatile unsigned long *)0xFEC80010)
+#define CursorControlReg_H_B	    ((volatile unsigned long *)0xFEC80014)
 
-#define LineInterrupt0_B		    ((volatile uint16_t *)0xFEC80018)
-#define LineInterrupt1_B		    ((volatile uint16_t *)0xFEC8001A)
-#define LineInterrupt2_B		    ((volatile uint16_t *)0xFEC8001C)
-#define LineInterrupt3_B		    ((volatile uint16_t *)0xFEC8001E)
+#define LineInterrupt0_B		    ((volatile unsigned long *)0xFEC80018)
+#define LineInterrupt1_B		    ((volatile unsigned long *)0xFEC8001A)
+#define LineInterrupt2_B		    ((volatile unsigned long *)0xFEC8001C)
+#define LineInterrupt3_B		    ((volatile unsigned long *)0xFEC8001E)
 
-#define MousePointer_Mem_B		    ((volatile uint16_t *)0xFEC80400)
-#define MousePtr_B_CTRL_Reg		    ((volatile uint16_t *)0xFEC80C00)
+#define MousePointer_Mem_B		    ((volatile unsigned long *)0xFEC80400)
+#define MousePtr_B_CTRL_Reg		    ((volatile unsigned long *)0xFEC80C00)
 
-#define MousePtr_B_X_Pos		    ((volatile uint16_t *)0xFEC80C02)
-#define MousePtr_B_Y_Pos		    ((volatile uint16_t *)0xFEC80C04)
-#define MousePtr_B_Mouse0		    ((volatile uint16_t *)0xFEC80C0A)
-#define MousePtr_B_Mouse1		    ((volatile uint16_t *)0xFEC80C0C)
-#define MousePtr_B_Mouse2		    ((volatile uint16_t *)0xFEC80C0E)
+#define MousePtr_B_X_Pos		    ((volatile unsigned long *)0xFEC80C02)
+#define MousePtr_B_Y_Pos		    ((volatile unsigned long *)0xFEC80C04)
+#define MousePtr_B_Mouse0		    ((volatile unsigned long *)0xFEC80C0A)
+#define MousePtr_B_Mouse1		    ((volatile unsigned long *)0xFEC80C0C)
+#define MousePtr_B_Mouse2		    ((volatile unsigned long *)0xFEC80C0E)
 
 #define ScreenText_B			    ((volatile char *)0xFECA0000)		    /* Text matrix */
 #define ColorText_B				    ((volatile uint8_t *)0xFECA8000)	    /* Color matrix */
 #define FG_CLUT_B 				    ((volatile unsigned long *)0xFECAC400)	/* Foreground LUT */
 #define BG_CLUT_B 				    ((volatile unsigned long *)0xFECAC440)	/* Background LUT */
 
-#define BM0_Control_Reg			    ((volatile uint16_t *)0xFEC80100)
-#define BM0_Addy_Pointer_Reg	    ((volatile uint16_t *)0xFEC80104)
+#define BM0_Control_Reg			    ((volatile unsigned long *)0xFEC80100)
+#define BM0_Addy_Pointer_Reg	    ((volatile unsigned long *)0xFEC80104)
 
-#define Sprite_0_CTRL			    ((volatile uint16_t *)0xFEC81000)
-#define Sprite_0_ADDY_HI		    ((volatile uint16_t *)0xFEC81002)
-#define Sprite_0_POS_X	     	    ((volatile uint16_t *)0xFEC81004)
-#define Sprite_0_POS_Y	   		    ((volatile uint16_t *)0xFEC81006)
+#define Sprite_0_CTRL			    ((volatile unsigned long *)0xFEC81000)
+#define Sprite_0_ADDY_HI		    ((volatile unsigned long *)0xFEC81002)
+#define Sprite_0_POS_X	     	    ((volatile unsigned long *)0xFEC81004)
+#define Sprite_0_POS_Y	   		    ((volatile unsigned long *)0xFEC81006)
 
 /*
  * Color lookup tables
